moved loading component to components directory
