+++ resources/js/pages/principal/schedules/juniorHigh/grade_9.vue
@@ -15,7 +15,7 @@
 </template>
 
 <script>
-import Loading from '../../../../components/principal/Loading';
+import Loading from './../../../../components/Loading';
 import ScheduleCard from '../../../../components/principal/ScheduleCard';
 
 export default {
